NIFI-8113 Minor adjustments based on review: formatting and moving QuestDB storages into a separate package
