+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-framework-core/src/main/java/org/apache/nifi/controller/status/history/storage/BufferedWriterFlushWorker.java
@@ -34,7 +34,7 @@ public class BufferedWriterFlushWorker implements Runnable {
     @Override
     public void run() {
         try {
-            bufferedWriterList.forEach(statusRepository -> statusRepository.flush());
+            bufferedWriterList.forEach(bufferedWriter -> bufferedWriter.flush());
         } catch (final Exception e) {
             LOGGER.error("Error happened during calling flush.", e);
         }
