+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-framework-core/src/main/java/org/apache/nifi/controller/status/history/storage/questdb/QuestDbComponentStatusStorage.java
@@ -14,7 +14,7 @@
  * See the License for the specific language governing permissions and
  * limitations under the License.
  */
-package org.apache.nifi.controller.status.history.storage;
+package org.apache.nifi.controller.status.history.storage.questdb;
 
 import io.questdb.cairo.sql.Record;
 import org.apache.commons.lang3.time.DateUtils;
@@ -29,6 +29,8 @@ import org.apache.nifi.controller.status.history.questdb.QuestDbContext;
 import org.apache.nifi.controller.status.history.questdb.QuestDbEntityReadingTemplate;
 import org.apache.nifi.controller.status.history.questdb.QuestDbEntityWritingTemplate;
 import org.apache.nifi.controller.status.history.questdb.QuestDbStatusSnapshotMapper;
+import org.apache.nifi.controller.status.history.storage.ComponentStatusStorage;
+import org.apache.nifi.controller.status.history.storage.StatusStorage;
 
 import java.util.Arrays;
 import java.util.Collections;
