+++ nifi-nar-bundles/nifi-framework-bundle/nifi-framework/nifi-resources/src/main/resources/conf/nifi.properties
@@ -128,9 +128,9 @@ nifi.components.status.repository.buffer.size=${nifi.components.status.repositor
 nifi.components.status.snapshot.frequency=${nifi.components.status.snapshot.frequency}
 
 # QuestDB Status History Repository Properties
-# nifi.status.repository.questdb.persist.node.days=${nifi.status.repository.questdb.persist.node.days}
-# nifi.status.repository.questdb.persist.component.days=${nifi.status.repository.questdb.persist.component.days}
-# nifi.status.repository.questdb.persist.location=${nifi.status.repository.questdb.persist.location}
+nifi.status.repository.questdb.persist.node.days=${nifi.status.repository.questdb.persist.node.days}
+nifi.status.repository.questdb.persist.component.days=${nifi.status.repository.questdb.persist.component.days}
+nifi.status.repository.questdb.persist.location=${nifi.status.repository.questdb.persist.location}
 
 # Site to Site properties
 nifi.remote.input.host=
